requirements.txt
fastapi
uvicorn[standard]
httpx
pyvips
Pillow
beautifulsoup4
fal-client
requests
python-dotenv
playwright
instagrapi
openai
anthropic
